New feature: AntiPhish.
If the master password is not saved, it would get queried in a dialog. However,
that dialog is just a prompt(). To make sure to users that it is genuine and not
a prompt from the website, we now display an unique random string in it that the
user can compare to what the extension options show.